Fix compiler and linker warnings
- Add type casts where C++20 deprecates implicit enum/float/int
  conversions (rays color_space, rays-video video, reflex shape/style)
- Cast NSImageRep* to NSBitmapImageRep* in rays bitmap.mm loader
- Remove unused const get_data overload for WallShape in reflex
- Wrap not-implemented stubs (SoundPlayer::rewind, vibrate,
  Timers::find_timers) with #pragma to suppress -Wmissing-noreturn
- xot extconf: only add -lstdc++ on gcc; clang++ auto-links libc++
  and the explicit -lc++ produced an ld duplicate-library warning


[[7A053F39-54F2-4D0A-AF3F-814E0C8DE517]]

+++ src/style.cpp
@@ -321,7 +321,7 @@ namespace Reflex
 
 		bool set_flow (Flow main, Flow sub)
 		{
-			return flow.set((main & FLOW_MASK) | ((sub & FLOW_MASK) << FLOW_SHIFT));
+			return flow.set(((int) main & FLOW_MASK) | (((int) sub & FLOW_MASK) << FLOW_SHIFT));
 		}
 
 		Flow flow_main () const
